face login register component loaded problem
