+++ src/app/(auth)/login/page.js
@@ -10,6 +10,8 @@ import * as yup from "yup";
 import { signIn } from "next-auth/react";
 import { toast } from "react-toastify";
 import { useRouter, useSearchParams } from "next/navigation";
+import Login from "@/app/components/Login/Login";
+import Register from "@/app/components/Register/Register";
 
 const defaultValues = {
   identifier: "web.kawsarahmed@gmail.com",
@@ -27,7 +29,7 @@ const schema = yup
   })
   .required();
 
-const Login = () => {
+const page = () => {
   const router = useRouter();
   const searchParams = useSearchParams();
   const callbackUrl = searchParams.get("callbackUrl");
@@ -151,236 +153,8 @@ const Login = () => {
               id="myTabContent"
               data-aos="fade-up"
             >
-              <div
-                className={`tab-pane fade ${
-                  isActive === "login" ? "active show" : ""
-                }`}
-                id="projects__one"
-                role="tabpanel"
-                aria-labelledby="projects__one"
-              >
-                <div className="col-xl-8 col-md-8 offset-md-2">
-                  <div className="loginarea__wraper">
-                    <div className="login__heading">
-                      <h5 className="login__title">Login</h5>
-                      <p className="login__description">
-                        Don't have an account yet?{" "}
-                        <a
-                          href="#"
-                          data-bs-toggle="modal"
-                          data-bs-target="#registerModal"
-                        >
-                          Sign up for free
-                        </a>
-                      </p>
-                    </div>
-
-                    <form onSubmit={handleSubmit(onSubmit)}>
-                      <div className="login__form">
-                        <label className="form__label">Username or email</label>
-                        <input
-                          className="common__login__input"
-                          type="text"
-                          placeholder="Your username or email"
-                          {...register("identifier")}
-                          value={defaultValues.identifier}
-                        />
-
-                        {errors?.identifier?.message && (
-                          <div
-                            class="invalid-feedback"
-                            style={{ display: "block" }}
-                          >
-                            {errors?.identifier?.message}
-                          </div>
-                        )}
-                      </div>
-                      <div className="login__form">
-                        <label className="form__label">Password</label>
-                        <input
-                          className="common__login__input"
-                          type="password"
-                          placeholder="Password"
-                          {...register("password")}
-                          value={defaultValues.password}
-                        />
-
-                        {errors?.password?.message && (
-                          <div
-                            class="invalid-feedback"
-                            style={{ display: "block" }}
-                          >
-                            {errors?.password?.message}
-                          </div>
-                        )}
-                      </div>
-                      <div className="login__form d-flex justify-content-between flex-wrap gap-2">
-                        <div className="form__check">
-                          <input id="forgot" type="checkbox" />
-                          <label htmlFor="forgot"> Remember me</label>
-                        </div>
-                        <div className="text-end login__form__link">
-                          <a href="#">Forgot your password?</a>
-                        </div>
-                      </div>
-                      <div className="login__button">
-                        <button
-                          className="default__button"
-                          type="submit"
-                          disabled={isLoading}
-                        >
-                          Log In
-                        </button>
-                      </div>
-                    </form>
-                    <div className="login__social__option">
-                      <p>or Log-in with</p>
-
-                      <ul className="login__social__btn">
-                        <li>
-                          <a
-                            className="default__button login__button__1 d-flex align-items-center"
-                            href="#"
-                          >
-                            <FaFacebookF className="mr-1" /> Facebook
-                          </a>
-                        </li>
-                        {/* <li>
-                          <a
-                            className="default__button d-flex align-items-center"
-                            href="#"
-                          >
-                            <FaGooglePlusG className="mr-1" /> Google
-                          </a>
-                        </li> */}
-
-                        <li>
-                          <button
-                            className="default__button login__button__1 d-flex align-items-center"
-                            onClick={handleGithubLogin}
-                          >
-                            <FaGithub className="mr-2" /> Github
-                          </button>
-                        </li>
-                      </ul>
-                    </div>
-                  </div>
-                </div>
-              </div>
-
-              <div
-                className={`tab-pane fade ${
-                  isActive === "register" ? "active show" : ""
-                }`}
-                id="projects__two"
-                role="tabpanel"
-                aria-labelledby="projects__two"
-              >
-                <div className="col-xl-8 offset-md-2">
-                  <div className="loginarea__wraper">
-                    <div className="login__heading">
-                      <h5 className="login__title">Register</h5>
-                      <p className="login__description">
-                        Already have an account?{" "}
-                        <button
-                          onClick={() => handelModal("login")}
-                          data-bs-toggle="modal"
-                          data-bs-target="#registerModal"
-                        >
-                          Log In
-                        </button>
-                      </p>
-                    </div>
-
-                    <form onSubmit={handleSubmit(onSubmit)}>
-                      <div className="row">
-                        <div className="col-xl-6">
-                          <div className="login__form">
-                            <label className="form__label">First Name</label>
-                            <input
-                              className="common__login__input"
-                              type="text"
-                              placeholder="First Name"
-                              {...register("firstName")}
-                            />
-                          </div>
-                        </div>
-                        <div className="col-xl-6">
-                          <div className="login__form">
-                            <label className="form__label">Last Name</label>
-                            <input
-                              className="common__login__input"
-                              type="password"
-                              placeholder="Last Name"
-                              {...register("lastName")}
-                            />
-                          </div>
-                        </div>
-                        <div className="col-xl-6">
-                          <div className="login__form">
-                            <label className="form__label">Username</label>
-                            <input
-                              className="common__login__input"
-                              type="password"
-                              placeholder="Username"
-                              {...register("userName")}
-                            />
-                          </div>
-                        </div>
-                        <div className="col-xl-6">
-                          <div className="login__form">
-                            <label className="form__label">Email</label>
-                            <input
-                              className="common__login__input"
-                              type="password"
-                              placeholder="Your Email"
-                              {...register("email")}
-                            />
-                          </div>
-                        </div>
-                        <div className="col-xl-6">
-                          <div className="login__form">
-                            <label className="form__label">Password</label>
-                            <input
-                              className="common__login__input"
-                              type="password"
-                              placeholder="Password"
-                              {...register("password")}
-                            />
-                          </div>
-                        </div>
-                        <div className="col-xl-6">
-                          <div className="login__form">
-                            <label className="form__label">
-                              Re-Enter Password
-                            </label>
-                            <input
-                              className="common__login__input"
-                              type="password"
-                              placeholder="Re-Enter Password"
-                              {...register("confirmPassword")}
-                            />
-                          </div>
-                        </div>
-                      </div>
-
-                      <div className="login__form d-flex justify-content-between flex-wrap gap-2">
-                        <div className="form__check">
-                          <input id="accept_pp" type="checkbox" />{" "}
-                          <label htmlFor="accept_pp">
-                            Accept the Terms and Privacy Policy
-                          </label>
-                        </div>
-                      </div>
-                      <div className="login__button">
-                        <a className="default__button" href="#">
-                          Register
-                        </a>
-                      </div>
-                    </form>
-                  </div>
-                </div>
-              </div>
+              <Login />
+              <Register />
             </div>
           </div>
 
@@ -412,4 +186,4 @@ const Login = () => {
   );
 };
 
-export default Login;
+export default page;
